tests/wpt/web-platform-tests/css/css-masking/parsing/mask-type-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: getComputedStyle().maskType</title>
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-mask-type">
<meta name="assert" content="mask-type computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("mask-type", "luminance");
test_computed_value("mask-type", "alpha");
</script>
</body>
</html>
